requirements.txt
absl-py==0.1.11
asn1crypto==0.24.0
astor==0.6.2
attrs==17.4.0
Automat==0.6.0
baidu-aip==2.2.0.0
bcrypt==3.1.4
beautifulsoup4==4.6.0
bleach==1.5.0
boto==2.48.0
boto3==1.5.22
botocore==1.8.36
bz2file==0.98
cchardet==2.1.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
coverage==4.5.1
cryptography==2.2.2
cssselect==1.0.3
cycler==0.10.0
decorator==4.2.1
demjson==2.2.4
docutils==0.14
downloader==0.98
editdistance==0.4
gast==0.2.0
gensim==3.2.0
geoip2==2.8.0
getproxy==0.2.3
gevent==1.2.2
greenlet==0.4.13
grpcio==1.10.0
gscholar==1.6.1
hanziconv==0.3.2
html5lib==0.9999999
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
ipykernel==4.8.0
ipython==4.2.1
ipython-genutils==0.2.0
isodate==0.6.0
jieba==0.39
jmespath==0.9.3
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==1.2.2
lxml==4.2.1
Markdown==2.6.11
matplotlib==2.1.2
maxminddb==1.3.0
mock==2.0.0
mysql-connector-python-rf==2.2.2
networkx==2.1
nltk==3.2.5
nose==1.3.7
numpy==1.14.0
pandas==0.22.0
paramiko==2.4.0
parsel==1.4.0
pbr==4.0.2
pdfminer.six==20170720
pexpect==4.3.1
pickleshare==0.7.4
Pillow==5.0.0
protobuf==3.5.1
psutil==5.4.3
ptyprocess==0.5.2
py-web-search==0.2.4
py2neo==3.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
pycryptodome==3.5.1
PyDispatcher==2.0.5
pyltp==0.2.1
pymongo==3.6.0
PyNaCl==1.2.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyperclip==1.6.0
pyquery==1.4.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
pyzmq==16.0.4
queuelib==1.5.0
rdflib==4.2.2
requests==2.18.4
retrying==1.3.3
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
Scrapy==1.5.0
selenium==3.10.0
service-identity==17.0.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.5.6
snownlp==0.12.3
SQLAlchemy==1.2.4
sshtunnel==0.1.2
stanfordcorenlp==3.7.0.2
tensorboard==1.6.0
tensorflow==1.6.0
termcolor==1.1.0
textrank==0.1.0
textrank4zh==0.3
Theano==1.0.1
tornado==4.5.3
tqdm==4.19.5
traitlets==4.3.2
Twisted==17.9.0
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
xlrd==1.1.0
zope.interface==4.4.3
